requirements.txt
PySide6
pyautogui
pynput
opencv-python
pillow
packaging
pyinstaller
google-genai
